+++ app.py
@@ -12,7 +12,7 @@ from application import db, app, login_manager
 from application.admin.routes import admin_bp
 from application.user.routes import user_bp
 from application.auth.routes import bp as auth_bp
-from application.add_meal_v2.routes import bp as add_meal_v2_bp
+from application.add_meal.routes import bp as add_meal_bp
 from typing import Optional
 
 # Config
@@ -30,12 +30,17 @@ def load_user(user_id: int):
 app.register_blueprint(admin_bp)
 app.register_blueprint(user_bp)
 app.register_blueprint(auth_bp)
-app.register_blueprint(add_meal_v2_bp)
+app.register_blueprint(add_meal_bp)
+
+
+# @app.errorhandler(404)
+# def page_not_found(e):
+#    return redirect("/")
 
 
 # Routes
 def default_return(next_page: Optional[str] = None):
-    return redirect(url_for("user.daily_log2"))
+    return redirect(url_for("user.daily_log"))
     if next_page:
         return redirect(next_page)
     if current_user.is_admin:
