+++ application/add_meal/templates/step4.html
@@ -1,7 +1,6 @@
 {% extends "base.html" %}
 
 {% block content %}
-<p>{{ tod }}</p>
 <p>{{ item.name }}</p>
 
 <form method="POST">
