+++ application/templates/base.html
@@ -23,7 +23,7 @@
                 <div class="d-flex w-100">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('user.daily_log2') }}">Daily Log (new)</a>
+                            <a class="nav-link" href="{{ url_for('user.daily_log') }}">Daily Log (new)</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link" href="{{ url_for('user.dashboard') }}">Dashboard</a>
