+++ application/user/templates/daily_log.html
@@ -50,7 +50,7 @@
         </a>
 
         <!-- Center Button (highlighted) -->
-        <a id="set_link_date" href="{{ url_for('add_meal_v2.get_barcode') }}"
+        <a id="set_link_date" href="{{ url_for('add_meal.find_item') }}"
             class="btn btn-success flex-fill mx-2 fw-bold rounded-pill">
             ＋ Add Item
         </a>
